requirements.txt
streamlit
langchain-community
openai
python-dotenv
